Effect를 별도로 분리하고 컴포넌트 순서 정리
